autherisation done

+++ controllers/comment.controller.js
@@ -1,11 +1,16 @@
 const commentModel = require("../db/models").UserComment
  
-const getComment= async (req ,res) =>{ 
- try{ const getComments= await commentModel.find().exec()
+const getComment= async (req ,res) =>{
+
+ try{  
+  //const userId = req.loggedInUser._id.toString()
+  //console.log(userId)
+  const getComments= await commentModel.find({}).exec()
     res.send(getComments)
+    //console.log(loggedInUser)
     return
 }catch (err){
-  res.status(500).send("sever not found")
+  res.status(500).send({message:"server not found"})
 }
 }
 const getOneComment = async (req, res) => {
@@ -20,6 +25,7 @@ const getOneComment = async (req, res) => {
   const postComment= async (req,res)=>{
     try{
         const {title}= req.body
+        //const userId = req.loggedInUser._id
         console.log(title)
         const newComm= await commentModel.create(title)
         res.send(newComm)
@@ -32,13 +38,23 @@ const getOneComment = async (req, res) => {
     try {
       const { _id } = req.params
       const {title } = req.body
-      if (!_id) res.status(404).send('Not Found')
+      if (!_id) {res.status(404).send({message:'Not Found'})
+      return
+    }
+      /*
+      const commentInfo = await commentModel.findById(_id).exec()
+      if (req.loggedInUser._id.toString() !== commentInfo?.userId?.toString()) {
+        res
+          .status(401)
+          .send({ errorMsg: 'you do not have access to edit this comment' })
+        return
+      }*/
   
       await commentModel.updateOne({ _id }, {title }).exec()
       const updateComment= await commentModel.findById(_id).exec()
       res.send(updateComment)
     } catch (err) {
-      res.status(500).send('Server error')
+      res.status(500).send({message:'Server error'})
     }
   }
   
@@ -46,6 +62,15 @@ const getOneComment = async (req, res) => {
     try{
       const {_id}= req.params
       if(!_id) res.status(404).send("ID not Found")
+      /*const commentInfo = await commentModel.findById(_id).exec()
+      if (req.loggedInUser._id.toString() !== commentInfo?.userId?.toString()) {
+        res
+          .status(401)
+          .send({ errorMsg: 'you do not have access to delete this comment' })
+        return
+      }*/
+  
+
       await commentModel.findByIdAndDelete(_id).exec()
       res.send ({message:`Comment with ID ${_id} has been deleted.`})
     }catch (err){
